Made pauze menu first version
