--- Assets/Scripts/UI/EndOfLevelUI.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c76ccd95081405a48a65ce3bf7d25f30
